clean up code and slight reorg
