+++ frameworks/Rust/faf/src/const_http.rs
@@ -1,25 +1,14 @@
-pub const fn const_len<T>(con: &[T]) -> usize {
-   con.len()
-}
+#![allow(dead_code)]
 
-pub const EPOLL_CTL_ADD: i32 = 1;
-pub const EPOLL_CTL_MOD: i32 = 3;
-pub const EPOLL_CTL_DEL: i32 = 2;
+use crate::util::const_len;
 
-pub const EPOLLIN: u32 = 0x1;
-pub const EPOLLPRI: u32 = 0x2;
-pub const EPOLLOUT: u32 = 0x4;
-pub const EPOLLRDNORM: u32 = 0x40;
-pub const EPOLLRDBAND: u32 = 0x80;
-pub const EPOLLWRNORM: u32 = 0x100;
-pub const EPOLLWRBAND: u32 = 0x200;
-pub const EPOLLMSG: u32 = 0x400;
-pub const EPOLLERR: u32 = 0x8;
-pub const EPOLLHUP: u32 = 0x10;
-pub const EPOLLET: u32 = 0x80000000;
-pub const EPOLLRDHUP: u32 = 0x2000;
-pub const EPOLLEXCLUSIVE: u32 = 0x10000000;
-pub const EPOLLONESHOT: u32 = 0x40000000;
+pub const GET: &[u8] = b"GET";
+pub const GET_LEN: usize = const_len(GET);
+
+pub const HTTP11: &[u8] = b"HTTP/1.1 200 OK";
+pub const SERVER: &[u8] = b"Server: F";
+pub const CRLF: &[u8] = b"\r\n";
+pub const CRLFCRLF: &[u8] = b"\r\n\r\n";
 
 pub const EXAMPLE_HTTP_RESPONSE: &[u8] = b"HTTP/1.1 200 OK\r\nContent-Length: 15\r\nContent-Type: text/plain\r\nServer: F\r\nDate: Wed, 24 Feb 2021 12:00:00 GMT\r\n\r\nHello, World!";
 pub const EXAMPLE_HTTP_RESPONSE_LEN: usize = const_len(EXAMPLE_HTTP_RESPONSE);
