+++ frameworks/Rust/faf/src/const_sys.rs
@@ -1,3 +1,5 @@
+#![allow(dead_code)]
+
 // Manually transformed from /usr/include/x86_64-linux-gnu/asm/unistd_64.h
 
 pub const SYS_READ: u32 = 0;
@@ -612,3 +614,24 @@ pub const INADDR_LOOPBACK: u32 = 2130706433;
 pub const INADDR_ANY: u32 = 0;
 pub const INADDR_BROADCAST: u32 = 4294967295;
 pub const INADDR_NONE: u32 = 4294967295;
+
+// EPOLL
+
+pub const EPOLL_CTL_ADD: i32 = 1;
+pub const EPOLL_CTL_MOD: i32 = 3;
+pub const EPOLL_CTL_DEL: i32 = 2;
+
+pub const EPOLLIN: u32 = 0x1;
+pub const EPOLLPRI: u32 = 0x2;
+pub const EPOLLOUT: u32 = 0x4;
+pub const EPOLLRDNORM: u32 = 0x40;
+pub const EPOLLRDBAND: u32 = 0x80;
+pub const EPOLLWRNORM: u32 = 0x100;
+pub const EPOLLWRBAND: u32 = 0x200;
+pub const EPOLLMSG: u32 = 0x400;
+pub const EPOLLERR: u32 = 0x8;
+pub const EPOLLHUP: u32 = 0x10;
+pub const EPOLLET: u32 = 0x80000000;
+pub const EPOLLRDHUP: u32 = 0x2000;
+pub const EPOLLEXCLUSIVE: u32 = 0x10000000;
+pub const EPOLLONESHOT: u32 = 0x40000000;
